implementação exceptions delete

+++ src/main/java/com/testesjava/curso_jpa_spring/services/UserService.java
@@ -2,8 +2,11 @@ package com.testesjava.curso_jpa_spring.services;
 
 import com.testesjava.curso_jpa_spring.entities.User;
 import com.testesjava.curso_jpa_spring.repositories.UserRepository;
+import com.testesjava.curso_jpa_spring.services.exceptions.DatabaseException;
 import com.testesjava.curso_jpa_spring.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +31,13 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
